Fixed Ray sprite + animations ; Refacto scripts

+++ src/RGB Unity/Assets/Scripts/RayStateMachine.cs
@@ -7,6 +7,6 @@ public class RayStateMachine : StateMachineBehaviour
     // OnStateExit is called before OnStateExit is called on any state inside this state machine
     public override void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        animator.SetFloat(RandomHaloAnimator, Random.Range(0,3));
+        animator.SetInteger(RandomHaloAnimator, Random.Range(0,3));
     }
 }
